backend/requirements.txt
mediapipe>=0.10.30
opencv-python>=4.8.0
numpy>=1.24.0
anthropic>=0.40.0
openai>=1.50.0
elevenlabs>=1.10.0
flask>=2.0.0
flask-cors>=4.0.0
pillow>=9.0.0
